Generic test runner: Migrate MUI V5 Slider (10/n)
[ghstack-poisoned]
