+++ packages/component-driver-mui-v5-test/src/examples/index.ts
@@ -3,6 +3,6 @@ export * from './checkbox';
 export * from './input';
 export * from './rating';
 export * from './select';
-export * from './Slider.examples';
+export * from './slider';
 export * from './Switch.examples';
 export * from './TextField.examples';
